backend/pyproject.toml
[tool.poetry]
name = "instanssi.org"
version = "1.0.0"
description = "Instanssi website"
authors = ["Tuomas Virtanen <katajakasa@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-countries = "^7.4.2"
django-crispy-forms = "^1.14.0"
django-imagekit = "^4.1.0"
django-twitter-tag = "^1.2.1"
django-classy-tags = "^4.0.0"
django-filter = "^22.1"
django-compressor = "^4.1"
django-qr-code = "^3.1.1"
django-auditlog = "^2.2.1"
celery = "^5.2.7"
Pillow = "^9.3.0"
social-auth-core = "^4.3.0"
social-auth-app-django = "^5.0.0"
django-libsass = "^0.9"
requests = "^2.28.1"
sentry-sdk = "^1.11.1"
django-redis = "^5.2.0"
djangorestframework = "^3.14.0"
django-oauth-toolkit = "^2.2.0"
requests-oauthlib = "^1.3.1"
tabulate = "^0.9.0"
arrow = "^1.2.3"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
uvicorn = "^0.20.0"
unidecode = "^1.3.6"
yarl = "^1.8.2"
orjson = "^3.8.5"

[tool.poetry.group.dev.dependencies]
Faker = "^15.3.3"
black = "^22.10.0"
isort = "^5.10.1"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.1.0"
freezegun = "^1.2.2"
requests-mock = "^1.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 109
target-version = ['py39']

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "Instanssi.test_settings"
FAIL_INVALID_TEMPLATE_VARS = true
minversion = "6.0"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
omit = [
    "**/migrations/*",
]